Okay this may break a bunch of things but I am on a branch. Have the base template for the new refactor need to build on it incrementally
