+++ app/add-workout.tsx
@@ -1,1019 +1,316 @@
-import { useRoute } from "@react-navigation/native";
-import { useNavigation, useRouter } from "expo-router";
-import { useCallback, useEffect, useLayoutEffect, useRef, useState } from "react";
-import { ActionSheetIOS, Alert, Keyboard, Platform, Text, TextInput, TouchableOpacity, TouchableWithoutFeedback, useColorScheme, View } from "react-native";
-
-import DraggableExercisePanel from "@/components/DraggableExercisePanel";
-import ExerciseInteractiveModal from "../components/ExerciseInteractiveModal";
-import { inferDefaultTags } from "../data/exerciseDefaultTags";
-import type { Exercise, ExerciseAction, ExerciseType, TagState } from "../types/workout";
-
-import useAccurateTimer from "@/hooks/useAccurateTimer";
-import { AntDesign, Ionicons } from "@expo/vector-icons";
-import AsyncStorage from "@react-native-async-storage/async-storage";
-import { useIsFocused } from "@react-navigation/native";
-import { nanoid } from 'nanoid/non-secure';
-import { Pressable } from "react-native";
-
-import { useWorkoutStore } from '../stores/workoutStore';
+import { useWorkoutStore } from '@/store/workoutStore';
+import type { WorkoutItem } from '@/types/workout';
+import { useEffect, useMemo, useState } from 'react';
+import {
+    Alert,
+    FlatList,
+    Modal,
+    Platform,
+    Pressable,
+    SafeAreaView,
+    ScrollView,
+    StyleSheet,
+    Text,
+    TextInput,
+    TouchableOpacity,
+    View,
+} from 'react-native';
 
 export default function AddWorkout() {
-    const route = useRoute();
-    const { mode = "live", templateId } = route.params || {};
-    const [editingTemplateId, setEditingTemplateId] = useState<number | string | null>(null);
-
-
-    // ─── NEW: keep the last modal duration in a ref ───
-    const editDurationRef = useRef(0);
-    const [editDuration, setEditDuration] = useState(0); // keep for UI/prop, but ref is source-of-truth
-    const [pendingFocusId, setPendingFocusId] = useState<string | null>(null);
-
-    // ───── Theme & Navigation ─────
-    const scheme = useColorScheme();
-    const navigation = useNavigation();
-    const router = useRouter();
-
-    const backgroundColor = scheme === "dark" ? "#000000" : "#ffffff";
-    const textColor = scheme === "dark" ? "#ffffff" : "#000000";
-    const borderColor = scheme === "dark" ? "#444" : "#ccc";
-    const cardColor = scheme === "dark" ? "#1e1e1e" : "#d1d1d1";
-
-    const inputBg = scheme === "dark" ? "#2a2a2a" : "#fff";
-    const dividerColor = scheme === "dark" ? "#333" : "#ccc";
-
-    // ───── Workout Info State ─────
-    const [date, setDate] = useState(new Date());
-    const [showDatePicker, setShowDatePicker] = useState(false);
-
-    // ───── Modal & Exercise State ─────
-    const [modalVisible, setModalVisible] = useState(false);
-    const [exerciseName, setExerciseName] = useState("");
-    const [exerciseNameBlurred, setExerciseNameBlurred] = useState(false);
-    const [lockedExerciseTitle, setLockedExerciseTitle] = useState("");
-    const [exerciseType, setExerciseType] = useState<ExerciseType>("unknown");;
-    const [tags, setTags] = useState<TagState>({});
-
-    const elapsedMs = useWorkoutStore((s) => s.elapsedMs);
-    const isActive = useWorkoutStore((s) => s.isActive);
-
-    // ───── Action Management State ─────
-    const [actionsList, setActionsList] = useState<ExerciseAction[]>([]);
-
-
-    const [editIndex, setEditIndex] = useState<number | null>(null);
-    const [resetExpansionTrigger, setResetExpansionTrigger] = useState(0);
-    const [triggerScrollToEnd, setTriggerScrollToEnd] = useState(false);
-
-    // const [elapsedTime, setElapsedTime] = useState(0);
-    // const timerRef = useRef<ReturnType<typeof setInterval> | null>(null);
-    const sessionKeyRef = useRef(`workoutTimer:${Date.now()}`);
-
-    const [totalApproxSeconds, setTotalApproxSeconds] = useState(0);
-    const [durationOverrideMin, setDurationOverrideMin] = useState<number | null>(null);
-    const [isDurationControlsOpen, setIsDurationControlsOpen] = useState(false);
-
-    const autoMinutes = Math.ceil((totalApproxSeconds ?? 0) / 60);
-    const currentMinutes = durationOverrideMin ?? autoMinutes;
-
-    const start = useWorkoutStore(s => s.start);
-    const stop = useWorkoutStore(s => s.stop);
-
-    // Live exercises from the unified store
-    const liveExercises = useWorkoutStore((s) => s.liveExercises);
-    const setAllLiveExercises = useWorkoutStore((s) => s.setAllLiveExercises);
-    const pushLiveExercise = useWorkoutStore((s) => s.pushLiveExercise);
-    const replaceLiveExerciseAt = useWorkoutStore((s) => s.replaceLiveExerciseAt);
-    const removeLiveExerciseAt = useWorkoutStore((s) => s.removeLiveExerciseAt);
-    const clearLiveExercises = useWorkoutStore((s) => s.clearLiveExercises);
-    const workoutName = useWorkoutStore(s => s.liveMeta.workoutName);
-    const preWorkoutNote = useWorkoutStore(s => s.liveMeta.preWorkoutNote);
-    const postWorkoutNote = useWorkoutStore(s => s.liveMeta.postWorkoutNote);
-
-    const setWorkoutName = useWorkoutStore(s => s.setWorkoutName);
-    const setPreWorkoutNote = useWorkoutStore(s => s.setPreNote);
-    const setPostWorkoutNote = useWorkoutStore(s => s.setPostNote);
-
-    // Clamp to [auto, 999]
-    const clampMinutes = (m: number) => Math.min(999, Math.max(autoMinutes, m));
-
-    const nudge = (delta: number) => {
-        setDurationOverrideMin(prev => {
-            const base = prev ?? autoMinutes;
-            return clampMinutes(base + delta);
-        });
-    };
-
-
-    // start global timer for live sessions
-    useEffect(() => {
-        if (mode === 'live' && !isActive) start();
-        return () => {
-            //if (mode === 'live') stop();
-        };
-    }, [mode, isActive, start]);
-
-
-    useEffect(() => {
-        const total = estimateWorkoutDurationSeconds(
-            (liveExercises ?? []).map(ex => ({
-                ...ex,
-                computedDurationInSeconds:
-                    typeof ex.computedDurationInSeconds === "number"
-                        ? ex.computedDurationInSeconds
-                        : Number(ex.computedDurationInSeconds ?? 0),
-            }))
-        );
-        // if user has an override, use that (minutes → seconds)
-        setTotalApproxSeconds(durationOverrideMin != null ? durationOverrideMin * 60 : total);
-    }, [liveExercises, durationOverrideMin]);
-
-    useEffect(() => {
-        const loadTemplate = async () => {
-            if (!templateId) return;
-
-            const stored = await AsyncStorage.getItem("savedTemplates");
-            if (!stored) return;
-
-            const parsed = JSON.parse(stored);
-            const found = parsed.find((t: any) => t.id.toString() === templateId.toString());
-            if (!found) return;
-
-            setAllLiveExercises(found.exercises);
-
-            if (mode === "live") {
-                setWorkoutName(`${found.name} Copy`);
-                start()
-            } else {
-                setWorkoutName(found.name);
-                setEditingTemplateId(found.id);
-                if (typeof found.durationOverrideMin === "number") {
-                    setDurationOverrideMin(found.durationOverrideMin);
-                } else if (typeof found.approxDurationInSeconds === "number") {
-                    setDurationOverrideMin(Math.ceil(found.approxDurationInSeconds / 60));
-                }
-            }
-        };
-
-        loadTemplate();
-    }, [mode, templateId]);
-
-    useEffect(() => {
-        return () => { stopWorkoutTimer(); };
-    }, []);
-
-    const isFocused = useIsFocused();
-    const {
-        displaySeconds: elapsedTime,
-        finalize: finalizeWorkoutTimer,
-        reset: resetWorkoutTimer,
-        persistNow: persistWorkoutTimer,
-        stop: stopWorkoutTimer,           // <-- add this from the hook
-    } = useAccurateTimer(sessionKeyRef.current, isFocused && mode === "live");
-
-
-    type Metrics = {
-        totalExercises: number;
-        totalSets: number;
-        totalWorkingSets: number;
-        approxDurationInSeconds: number;
-    };
-
-    function computeMetrics(exercises: Array<{ actions?: any[] }>): Omit<Metrics, "approxDurationInSeconds"> {
-        let totalSets = 0;
-        let totalWorkingSets = 0;
-
-        for (const ex of exercises ?? []) {
-            const sets = (ex.actions ?? []).filter(a => a?.type === "set");
-            totalSets += sets.length;
-            totalWorkingSets += sets.filter(s => !s?.isWarmup).length;
-        }
-        return {
-            totalExercises: exercises?.length ?? 0,
-            totalSets,
-            totalWorkingSets,
-        };
-    }
-    // ───── Workout Save (To Implement) ─────
-    const saveWorkout = useCallback(async () => {
-        const exs = liveExercises;
-        if (!workoutName || (exs?.length ?? 0) === 0) return;
-
-        const computedApprox = estimateWorkoutDurationSeconds(exs);
-
-        // clamp minutes if overridden
-        const approxMinutes =
-            durationOverrideMin != null
-                ? clampMinutes(durationOverrideMin) // [auto, 999]
-                : Math.ceil(computedApprox / 60);
-
-        const approxDurationInSeconds = approxMinutes * 60;
-
-        // build metrics
-        const baseCounts = computeMetrics(exs);
-        const metrics: Metrics = { ...baseCounts, approxDurationInSeconds };
-
-        // Build the base object (shared)
-        let workout: any = {
-            id: editingTemplateId ?? Date.now(),
-            name: workoutName,
-            exercises: exs,
-            approxDurationInSeconds,   // stays at top level for existing UI
-            metrics,                   // <— NEW: totals + approx
-            durationOverrideMin: durationOverrideMin,
-        };
-
-        let key = "savedWorkouts";
-
-        try {
-            if (mode === "template") {
-                key = "savedTemplates";
-                const existing = await AsyncStorage.getItem(key);
-                const list = existing ? JSON.parse(existing) : [];
-
-                if (editingTemplateId) {
-                    // UPDATE
-                    const idx = list.findIndex((t: any) => t.id.toString() === String(editingTemplateId));
-                    if (idx !== -1) {
-                        const prev = list[idx];
-                        list[idx] = {
-                            ...prev,
-                            ...workout,
-                            id: prev.id,
-                            createdOn: prev.createdOn ?? new Date().toISOString(),
-                            updatedOn: new Date().toISOString(),
-                            usageCount: prev.usageCount ?? 0,
-                        };
-
-                        console.log("Updating Template Payload:\n" + JSON.stringify(list[idx], null, 2));
-                    }
-                    await AsyncStorage.setItem(key, JSON.stringify(list, null, 2));
-                } else {
-                    // CREATE
-                    const newTemplate = {
-                        ...workout,
-                        createdBy: "Ishan Arefin",
-                        createdOn: new Date().toISOString(),
-                        usageCount: 0,
-                    };
-                    list.push(newTemplate);
-                    await AsyncStorage.setItem(key, JSON.stringify(list, null, 2));
-                }
-            } else {
-                // LIVE workout save
-                const existing = await AsyncStorage.getItem(key);
-                const list = existing ? JSON.parse(existing) : [];
-                const liveWorkout = {
-                    ...workout,
-                    preWorkoutNote,
-                    postWorkoutNote,
-                    date: date.toISOString(),
-                    actualDurationInSeconds: finalizeWorkoutTimer(),
-                    createdBy: "Ishan Arefin",
-                    completedAt: date.toISOString()
-                };
-                list.push(liveWorkout);
-                await AsyncStorage.setItem(key, JSON.stringify(list, null, 2));
-                clearLiveExercises();
-                stopWorkoutTimer();
-                stop();
-                setWorkoutName('');
-                setPreWorkoutNote('');
-                setPostWorkoutNote('');
-            }
-
-
-            navigation.goBack();
-        } catch (e) {
-            console.error("Failed to save workout:", e);
-        }
-    }, [mode, workoutName, preWorkoutNote, postWorkoutNote, date, liveExercises, editingTemplateId]);
-
-
-    const disabledColor = scheme === "dark" ? "#666" : "#bbb";
-    const canSave =
-        (workoutName?.trim().length ?? 0) > 0 &&
-        (liveExercises.length > 0);
-
-    function clock(ms: number) {
-        const s = Math.floor(ms / 1000);
-        const hh = String(Math.floor(s / 3600)).padStart(2, '0');
-        const mm = String(Math.floor((s % 3600) / 60)).padStart(2, '0');
-        const ss = String(s % 60).padStart(2, '0');
-        return `${hh}:${mm}:${ss}`;
-    }
-
-    // ───── Layout Effect for Header Buttons ─────
-    useLayoutEffect(() => {
-        navigation.setOptions({
-            headerTitle: () => {
-                if (mode === "live") {
-                    return (
-                        <View style={{ flexDirection: "row", alignItems: "center", paddingHorizontal: 4, flexShrink: 1 }}>
-                            <Ionicons name="time-outline" size={16} color={textColor} style={{ marginRight: 4 }} />
-                            <Text
-                                numberOfLines={1}
-                                style={{ fontSize: 16, fontWeight: "bold", color: textColor }}
-                            >
-                                {/* {formatElapsedTime(elapsedTime)} */}
-                                {isActive ? clock(elapsedMs) : clock(0)}
-                            </Text>
-                        </View>
-                    );
-                } else if (mode === "template" && !editingTemplateId) {
-                    return (
-                        <View style={{ flexDirection: "row", alignItems: "center", paddingHorizontal: 4, flexShrink: 1 }}>
-                            <AntDesign name="form" size={16} color={textColor} style={{ marginRight: 4 }} />
-                            <Text
-                                numberOfLines={1}
-                                style={{ fontSize: 16, fontWeight: "bold", color: textColor }}
-                            >
-                                Add Template
-                            </Text>
-                        </View>
-                    );
-                } else if (mode === "template" && editingTemplateId) {
-                    return (
-                        <View style={{ flexDirection: "row", alignItems: "center", paddingHorizontal: 4, flexShrink: 1 }}>
-                            <AntDesign name="form" size={16} color={textColor} style={{ marginRight: 4 }} />
-                            <Text numberOfLines={1} style={{ fontSize: 16, fontWeight: "bold", color: textColor }}>
-                                Edit Template
-                            </Text>
-                        </View>
-                    );
-                } else {
-                    return (
-                        <Text style={{ fontSize: 16, fontWeight: "bold", color: textColor }}>
-                            Add Workout
-                        </Text>
-                    );
-                }
-            },
-            headerRight: () => (
-                <TouchableOpacity
-                    disabled={!canSave}
-                    accessibilityRole="button"
-                    accessibilityState={{ disabled: !canSave }}
-                    onPress={() => {
-                        if (!canSave) return;
-                        if (mode === "live") {
-                            if (Platform.OS === "ios") {
-                                ActionSheetIOS.showActionSheetWithOptions(
-                                    {
-                                        title: "Finish Live Workout?",
-                                        message:
-                                            "Are you sure you want to finish this live workout? Once saved, it cannot be modified.",
-                                        options: ["Cancel", "Save Workout"],
-                                        cancelButtonIndex: 0,
-                                        destructiveButtonIndex: 1,
-                                        userInterfaceStyle: scheme === "dark" ? "dark" : "light",
-                                    },
-                                    (buttonIndex) => {
-                                        if (buttonIndex === 1) saveWorkout();
-                                    }
-                                );
-                            } else {
-                                Alert.alert(
-                                    "Finish Live Workout?",
-                                    "Are you sure you want to finish this live workout? Once saved, it cannot be modified.",
-                                    [
-                                        { text: "Cancel", style: "cancel" },
-                                        { text: "Save Workout", style: "destructive", onPress: saveWorkout },
-                                    ]
-                                );
-                            }
-                        } else {
-                            saveWorkout();
-                        }
-                    }}
-                >
-                    <Text
-                        style={{
-                            fontSize: 16,
-                            fontWeight: "bold",
-                            color: canSave ? textColor : disabledColor,
-                            opacity: canSave ? 1 : 0.6,
-                        }}
-                    >
-                        Done
-                    </Text>
-                </TouchableOpacity>
-            ),
-            headerLeft: () => (
-                <TouchableOpacity onPress={confirmClose}>
-                    <Text style={{ fontSize: 16, color: textColor }}>Cancel</Text>
-                </TouchableOpacity>
-            )
-        });
-    }, [navigation, textColor, saveWorkout, elapsedMs, isActive, mode, canSave, workoutName, liveExercises.length]);
-
-    // ───── Modal & Exercise Handlers ─────
-    const closeModal = () => {
-        setModalVisible(false);
-        setActionsList([]);
-        setExerciseName("");
-        setExerciseNameBlurred(false);
-        setLockedExerciseTitle("");
-        setEditIndex(null);
-        setResetExpansionTrigger(prev => prev + 1);
-        setTags({});
-    };
-
-    const confirmClose = () => {
-        const isWorkoutEmpty =
-            !workoutName?.trim() &&
-            !preWorkoutNote?.trim() &&
-            !postWorkoutNote?.trim() &&
-            (liveExercises.length === 0);             // <- use store array
-
-        const discardAndExit = async () => {
-            try { stopWorkoutTimer(); } catch { }
-            try { clearLiveExercises(); } catch { }    // <- clear the store
-            try { stop(); } catch { }                  // <- your live timer store stop()
-            try { setWorkoutName(''); setPreWorkoutNote(''); setPostWorkoutNote(''); } catch { }
-            navigation.goBack();
-        };
-
-        if (isWorkoutEmpty) {
-            // no dialog; just tear down
-            discardAndExit();
-            return;
-        }
-
-        if (Platform.OS === 'ios') {
-            ActionSheetIOS.showActionSheetWithOptions(
-                {
-                    message: 'Are you sure you want to discard this workout? Your current progress will be lost.',
-                    options: ['Cancel', 'Delete'],
-                    cancelButtonIndex: 0,
-                    destructiveButtonIndex: 1,
-                    userInterfaceStyle: 'dark',
-                },
-                async (buttonIndex) => {
-                    if (buttonIndex === 1) {
-                        await discardAndExit();             // <- clear store + draft + timer
-                    }
-                }
-            );
-        } else {
-            Alert.alert(
-                'Discard Workout?',
-                'Are you sure you want to discard this workout? This action cannot be undone.',
-                [
-                    { text: 'Cancel', style: 'cancel' },
-                    {
-                        text: 'Discard',
-                        style: 'destructive',
-                        onPress: () => { discardAndExit(); } // <- clear store + draft + timer
-                    },
-                ]
-            );
-        }
-    };
-
-    const confirmDeleteExercise = (indexToDelete: number) => {
-        if (Platform.OS === "ios") {
-            ActionSheetIOS.showActionSheetWithOptions(
-                {
-                    options: ["Cancel", "Delete Exercise"],
-                    destructiveButtonIndex: 1,
-                    cancelButtonIndex: 0,
-                    message: "Are you sure you want to delete this exercise from the workout? You will lose all of the data associated with it. This action cannot be undone.",
-                    userInterfaceStyle: "dark",
-                },
-                (buttonIndex) => {
-                    if (buttonIndex === 1) {
-                        handleDeleteExercise(indexToDelete);
-                    }
-                }
-            );
-        } else {
-            Alert.alert(
-                "Delete Exercise?",
-                "Are you sure you want to delete this exercise from the workout? You will lose all of the data associated with it. This action cannot be undone.",
-                [
-                    { text: "Cancel", style: "cancel" },
-                    { text: "Delete", style: "destructive", onPress: () => handleDeleteExercise(indexToDelete) }
-                ]
-            );
-        }
-    };
-
-    const handleSaveExercise = () => {
-        if (!exerciseName || actionsList.length === 0) return;
-
-        const computedDurationInSeconds = estimateExerciseDurationWorstCase(actionsList);
-
-        const newExercise: Exercise = {
-            name: exerciseName,
-            type: exerciseType,
-            actions: computeNumberedActions(actionsList),
-            editDurationInSeconds: editDurationRef.current,
-            computedDurationInSeconds,
-            tags,
-        };
-
-        if (editIndex !== null) {
-            replaceLiveExerciseAt(editIndex, newExercise);
-        } else {
-            pushLiveExercise(newExercise);
-        }
-
-        // ✅ reset modal state so you SEE the result in the list
-        closeModal();                   // hide modal
-        setEditIndex(null);             // so next open is “add” not “edit”
-        setExerciseName('');
-        setActionsList([]);             // start fresh
-        editDurationRef.current = 0;    // reset local edit timer if you use one
-    };
-
-    // ───── Action Handlers ─────
-    const addSet = () => {
-        const id = nanoid();
-        let newSet: any = {
-            id,
-            type: "set",
-        };
-
-        switch (exerciseType) {
-            case "weighted":
-                newSet = {
-                    ...newSet,
-                    weight: "",
-                    weightUnit: "lb",
-                    reps: "",
-                    unit: "lb",
-                    note: "",
-                    isWarmup: false,
-                    RPE: 0
-                };
-                break;
-            case "bodyweight":
-                newSet = {
-                    ...newSet,
-                    reps: "",
-                    unit: "",
-                    note: "",
-                    isWarmup: false,
-                    RPE: 0
-                };
-                break;
-            case "duration":
-                newSet = {
-                    ...newSet,
-                    value: "",
-                    unit: "sec"
-                };
-                break;
-            case "weighted duration":
-                newSet = {
-                    ...newSet,
-                    weight: "",
-                    weightUnit: "lb",
-                    value: "",
-                    valueUnit: "sec"
-                };
-                break;
-            case "weighted distance":
-                newSet = {
-                    ...newSet,
-                    weight: "",
-                    weightUnit: "lb",
-                    value: "",
-                    valueUnit: "yd"
-                };
-                break;
-            default:
-                newSet = {
-                    ...newSet,
-                    reps: "",
-                    unit: ""
-                };
-        }
-
-        const updatedList = computeNumberedActions([...actionsList, newSet]);
-        setPendingFocusId(id);
-        setActionsList(updatedList);
-        setTriggerScrollToEnd(true);
-        console.log("Updated Actions List:", JSON.stringify(updatedList, null, 2));
-    };
-
-    const addRest = () => {
-        const id = nanoid();
-        const newRest = {
-            id,
-            type: "rest",
-            value: "",
-            restInSeconds: 0
-        };
-
-        const updatedList = computeNumberedActions([...actionsList, newRest]);
-        setPendingFocusId(id);
-        setActionsList(updatedList);
-        setTriggerScrollToEnd(true);
-        console.log("Updated Actions List:", JSON.stringify(updatedList, null, 2));
-    };
-
-    const deleteActionById = (idToDelete: string) => {
-        const filtered = actionsList.filter(action => action.id !== idToDelete);
-        const updated = computeNumberedActions(filtered);
-        setActionsList(updated);
-
-        console.log(`Deleted action with ID ${idToDelete}`);
-        console.log("Updated Actions List:", JSON.stringify(updated, null, 2));
-    };
-    const updateActionValue = (
-        id: string,
-        field: "reps" | "weight" | "value" | "unit" | "weightUnit" | "valueUnit" | "note" | "isWarmup" | "RPE" | "restInSeconds",
-        value: string
-    ) => {
-        setActionsList(prev =>
-            prev.map(action =>
-                action.id === id ? { ...action, [field]: value } : action
-            )
-        );
-    };
-    const computeNumberedActions = (actions) => {
-        let setCount = 1;
-        let restCount = 1;
-
-        return actions.map((action) => {
-            const base = { ...action }; // preserve all fields including id
-
-            if (action.type === "set" && !action.isWarmup) {
-                return { ...base, setNumber: setCount++ };
-            } else if (action.type === "rest") {
-                return { ...base, restNumber: restCount++ };
-            } else {
-                return { ...base, setNumber: null };
-            }
-        });
-    };
-
-    const handleDeleteExercise = (indexToDelete: number) => {
-        removeLiveExerciseAt(indexToDelete);
-    };
-
-    // ───── Date Picker Handler ─────
-    const onChangeDate = (event: any, selectedDate?: Date) => {
-        setShowDatePicker(Platform.OS === "ios");
-        if (selectedDate) setDate(selectedDate);
-    };
-
-    function estimateExerciseDurationWorstCase(actions: ExerciseAction[]) {
-        const totalRest = actions
-            .filter(a => a.type === "rest")
-            .reduce((sum, a: any) => sum + Number(a.restInSeconds ?? 0), 0); // <-- Number()
-
-        const setCount = actions.filter(a => a.type === "set").length;
-
-        return totalRest + setCount * 70; // number
-    }
-
-    const INTER_EXERCISE_BUFFER_SEC = 90; // setup/log/water between exercises
-
-    function estimateWorkoutDurationSeconds(exercises: Exercise[]) {
-        if (!exercises?.length) return 0;
-
-        return exercises.reduce((total, ex, idx) => {
-            const perExercise =
-                typeof ex.computedDurationInSeconds === "number"
-                    ? ex.computedDurationInSeconds
-                    : estimateExerciseDurationWorstCase(ex.actions); // fallback for older items
-
-            const gap = idx > 0 ? INTER_EXERCISE_BUFFER_SEC : 0; // N-1 gaps
-            return total + Number(perExercise || 0) + gap;
-        }, 0);
-    }
-
-    function secondsToApproxMinutes(sec: number): string {
-        const minutes = Math.ceil((sec || 0) / 60);
-        return `${minutes} min${minutes !== 1 ? "s" : ""}`;
-    }
-
-    return (
-        <>
-            <TouchableWithoutFeedback onPress={Keyboard.dismiss} accessible={false}>
-                <View style={{ flex: 1, padding: 20 }}>
-                    {/* Motivational Quote: this could be made up of quotes from the user themselves after they keep tracking workouts*/}
-                    <Text style={{
-                        color: "#888",
-                        fontSize: 12,
-                        fontStyle: "italic",
-                        textAlign: "center",
-                        marginBottom: 6,
-                        paddingHorizontal: 12
-                    }}>
-                        "The worst thing I can be is the same as everybody else. I hate that."
-                        – Arnold Schwarzenegger
-                    </Text>
-                    {/* Workout Info */}
-                    <View style={{
-                        backgroundColor: cardColor,
-                        borderRadius: 12,
-                        padding: 12,
-                        marginBottom: 20,
-                        shadowColor: scheme === "dark" ? "#000" : "#aaa",
-                        shadowOpacity: scheme === "dark" ? 0.15 : 0.3,
-                        shadowRadius: 4,
-                        shadowOffset: { width: 0, height: 2 },
-                        elevation: 3,
-                        borderColor: scheme === "dark" ? "transparent" : "#ddd",
-                        borderWidth: scheme === "dark" ? 0 : 1,
-                    }}>
-                        {/* Pre-Workout Notes */}
-                        {mode === "live" && (
-                            <TextInput
-                                value={preWorkoutNote}
-                                onChangeText={setPreWorkoutNote}
-                                placeholder="Write a pre-workout note"
-                                placeholderTextColor={scheme === "dark" ? "#888" : "#aaa"}
-                                multiline
-                                scrollEnabled
-                                blurOnSubmit={false}
-                                style={{
-                                    color: textColor,
-                                    backgroundColor: inputBg,
-                                    borderRadius: 8,
-                                    paddingHorizontal: 10,
-                                    paddingTop: 12,
-                                    paddingBottom: 10,
-                                    height: 65,
-                                    fontSize: 14,
-                                    textAlignVertical: "top"
-                                }}
-                            />
-                        )}
-                        {mode === "live" && (
-                            <View style={{
-                                height: 1,
-                                backgroundColor: dividerColor,
-                                opacity: 0.4,
-                                marginVertical: 6
-                            }} />
-                        )}
-
-                        {/* Workout Name */}
-                        <TextInput
-                            value={workoutName}
-                            onChangeText={(text) => setWorkoutName(text)}
-                            placeholder="e.g., Push Day, Leg Day"
-                            placeholderTextColor={scheme === "dark" ? "#888" : "#aaa"}
-                            style={{
-                                color: textColor,
-                                backgroundColor: inputBg,
-                                borderRadius: 8,
-                                padding: 10,
-                                fontSize: 14
-                            }}
-                        />
-                        {mode === "live" && (
-                            <View style={{
-                                height: 1,
-                                backgroundColor: dividerColor,
-                                opacity: 0.4,
-                                marginVertical: 6
-                            }} />
-                        )}
-
-                        {mode === "live" && (
-                            <TextInput
-                                value={postWorkoutNote}
-                                onChangeText={setPostWorkoutNote}
-                                placeholder="Write a post-workout note"
-                                placeholderTextColor={scheme === "dark" ? "#888" : "#aaa"}
-                                multiline
-                                scrollEnabled
-                                blurOnSubmit={false}
-                                style={{
-                                    color: textColor,
-                                    backgroundColor: inputBg,
-                                    borderRadius: 8,
-                                    paddingHorizontal: 10,
-                                    paddingTop: 12,
-                                    paddingBottom: 10,
-                                    height: 65,
-                                    fontSize: 14,
-                                    textAlignVertical: "top"
-                                }}
-                            />
-                        )}
-
-
-                        <View style={{
-                            height: 1,
-                            backgroundColor: dividerColor,
-                            opacity: 0.4,
-                            marginVertical: 6
-                        }} />
-
-                        <View>
-                            <TouchableOpacity
-                                onPress={() => {
-                                    setExerciseName("");
-                                    setExerciseType("unknown");
-                                    setActionsList([]);
-                                    setLockedExerciseTitle("");
-                                    setExerciseNameBlurred(false);
-                                    setEditIndex(null);
-                                    setEditDuration(0);
-                                    editDurationRef.current = 0;
-                                    setTags({});
-                                    setModalVisible(true);
-                                }}
-                                style={{
-                                    backgroundColor: "#1e90ff",
-                                    borderRadius: 8,
-                                    paddingVertical: 12,
-                                    paddingHorizontal: 16,
-                                    alignItems: "center",
-                                }}
-                            >
-                                <Text style={{ color: "white", fontWeight: "bold", fontSize: 15 }}>
-                                    + Add Exercise
-                                </Text>
-                            </TouchableOpacity>
-
-
-                            {/* Approximate total workout time — templates only */}
-                            {mode === "template" && (liveExercises?.length ?? 0) > 0 && (
-                                <View style={{ marginTop: 8 }}>
-                                    <Text
-                                        style={{
-                                            fontSize: 12,
-                                            color: scheme === "dark" ? "#9ca3af" : "#4b5563",
-                                            marginBottom: 2,
-                                        }}
-                                    >
-                                        Estimated Workout Length:
-                                    </Text>
-
-                                    <View style={{ flexDirection: "row", alignItems: "center" }}>
-                                        {/* Time pill = toggle */}
-                                        <Pressable
-                                            onPress={() => setIsDurationControlsOpen(open => !open)}
-                                            style={{
-                                                paddingVertical: 5,
-                                                paddingHorizontal: 10,
-                                                borderRadius: 10,
-                                                borderWidth: 1,
-                                                borderColor: borderColor ?? "#d1d5db",
-                                                marginRight: 6,
-                                                minWidth: 96,
-                                                justifyContent: "center",
-                                                alignItems: "center",
-                                            }}
-                                        >
-                                            <Text
-                                                style={{
-                                                    fontSize: 15,
-                                                    fontWeight: "700",
-                                                    color: "#1e90ff",
-                                                    textAlign: "center",
-                                                    fontVariant: ["tabular-nums"],
-                                                }}
-                                            >
-                                                ~{currentMinutes} mins
-                                            </Text>
-                                        </Pressable>
-
-                                        {isDurationControlsOpen && (
-                                            <View style={{ flexDirection: "row", alignItems: "center" }}>
-                                                {[-5, -1, +1, +5].map((delta) => (
-                                                    <Pressable
-                                                        key={delta}
-                                                        onPress={() => nudge(delta)}
-                                                        style={{
-                                                            height: 30,
-                                                            minWidth: 32, // 👈 narrower
-                                                            paddingHorizontal: 4,
-                                                            borderRadius: 6,
-                                                            backgroundColor: "#1e90ff",
-                                                            marginRight: 4,
-                                                            justifyContent: "center",
-                                                            alignItems: "center",
-                                                        }}
-                                                    >
-                                                        <Text style={{ color: "white", fontWeight: "700", fontSize: 12 }}>
-                                                            {delta > 0 ? `+${delta}` : delta}
-                                                        </Text>
-                                                    </Pressable>
-                                                ))}
-
-                                                {/* Auto button */}
-                                                <Pressable
-                                                    onPress={() => setDurationOverrideMin(null)}
-                                                    style={{
-                                                        height: 30,
-                                                        paddingHorizontal: 6,
-                                                        borderRadius: 6,
-                                                        borderWidth: 1,
-                                                        borderColor: borderColor ?? "#d1d5db",
-                                                        marginRight: 4,
-                                                        justifyContent: "center",
-                                                        alignItems: "center",
-                                                    }}
-                                                >
-                                                    <Text
-                                                        style={{
-                                                            color: scheme === "dark" ? "#9ca3af" : "#6b7280",
-                                                            fontWeight: "700",
-                                                            fontSize: 12,
-                                                        }}
-                                                    >
-                                                        Auto
-                                                    </Text>
-                                                </Pressable>
-
-                                                {/* X button */}
-                                                <Pressable
-                                                    onPress={() => setIsDurationControlsOpen(false)}
-                                                    style={{
-                                                        height: 30,
-                                                        paddingHorizontal: 10,
-                                                        borderRadius: 6,
-                                                        backgroundColor: "#ef4444",
-                                                        justifyContent: "center",
-                                                        alignItems: "center",
-                                                    }}
-                                                >
-                                                    <Text style={{ color: "white", fontWeight: "700", fontSize: 12 }}>X</Text>
-                                                </Pressable>
-                                            </View>
-                                        )}
-                                    </View>
-                                </View>
-                            )}
-                        </View>
-                    </View>
-
-                    {/* Draggable Exercise Panel with List of Exercises */}
-                    {liveExercises.length > 0 && (
-                        <DraggableExercisePanel
-                            exercises={liveExercises}
-                            onPressExercise={(index) => {
-                                const exercise = liveExercises[index];
-                                setEditIndex(index);
-                                setExerciseName(exercise.name);
-                                setExerciseType(exercise.type);
-                                setActionsList(exercise.actions);
-                                setLockedExerciseTitle(exercise.name);
-                                setExerciseNameBlurred(true);
-                                const d = exercise.editDurationInSeconds || 0;
-                                setEditDuration(d);
-                                editDurationRef.current = d;
-                                setModalVisible(true);
-                                setTags(exercise.tags ?? {});
-                            }}
-                            onDeleteExercise={confirmDeleteExercise}
-                            mode={mode}
-                        />
-                    )}
-                </View>
-            </TouchableWithoutFeedback>
-
-            <ExerciseInteractiveModal
-                visible={modalVisible}
-                onClose={closeModal}
-                onSave={handleSaveExercise}
-                exerciseName={exerciseName}
-                exerciseNameBlurred={exerciseNameBlurred}
-                lockedExerciseTitle={lockedExerciseTitle}
-                exerciseType={exerciseType}
-                actionsList={actionsList}
-                updateActionValue={updateActionValue}
-                updateActionsList={setActionsList}
-                onSelectExercise={(exercise, type) => {
-                    setExerciseName(exercise);
-                    setExerciseType(type);
-                    setExerciseNameBlurred(true);
-                    setLockedExerciseTitle(exercise);
-                    setTags(inferDefaultTags(exercise, type));
-                }}
-                onChangeExerciseName={(text) => {
-                    setExerciseName(text);
-                    setExerciseNameBlurred(false);
-                }}
-                addSet={addSet}
-                addRest={addRest}
-                onDeleteAction={deleteActionById}
-                isEditing={editIndex !== null}
-                resetExpansionTrigger={resetExpansionTrigger}
-                scrollToBottom={triggerScrollToEnd}
-                onScrolledToBottom={() => setTriggerScrollToEnd(false)}
-                initialEditDuration={editDuration}
-                onCloseWithDuration={(duration) => {
-                    // keep both in sync
-                    setEditDuration(duration);
-                    editDurationRef.current = duration;
-                }}
-                trackTime={mode !== "template"}
-                mode={mode}
-                tags={tags}
-                onChangeTags={setTags}
-                pendingFocusId={pendingFocusId}
-                onPendingFocusHandled={() => setPendingFocusId(null)}
-            />
-        </>
-    );
+  const draft = useWorkoutStore((s) => s.draft);
+  const history = useWorkoutStore((s) => s.history);
+  const startDraft = useWorkoutStore((s) => s.startDraft);
+  const setDraftName = useWorkoutStore((s) => s.setDraftName);
+  const addNote = useWorkoutStore((s) => s.addNote);
+  const addExercise = useWorkoutStore((s) => s.addExercise);
+  const addCustom = useWorkoutStore((s) => s.addCustom);
+  const elapsedSeconds = useWorkoutStore((s) => s.elapsedSeconds);
+  const pause = useWorkoutStore((s) => (s as any).pause);
+  const resume = useWorkoutStore((s) => (s as any).resume);
+  // const finishAndSave = useWorkoutStore((s) => s.finishAndSave); // disabled for now
+
+  useEffect(() => {
+    if (!draft) startDraft('');
+  }, [draft, startDraft]);
+
+  // live tick
+  const [tick, setTick] = useState(0);
+  useEffect(() => {
+    const id = setInterval(() => setTick((t) => t + 1), 1000);
+    return () => clearInterval(id);
+  }, []);
+
+  const mmss = useMemo(() => {
+    const total = elapsedSeconds();
+    const mm = Math.floor(total / 60).toString().padStart(2, '0');
+    const ss = (total % 60).toString().padStart(2, '0');
+    return `${mm}:${ss}`;
+    // NOTE: depend on tick so the display updates every second
+  }, [elapsedSeconds, tick]);
+
+  const [exerciseOpen, setExerciseOpen] = useState(false);
+  const [noteOpen, setNoteOpen] = useState(false);
+  const [customOpen, setCustomOpen] = useState(false);
+  const [finishOpen, setFinishOpen] = useState(false);
+
+  const [exerciseName, setExerciseName] = useState('');
+  const [noteText, setNoteText] = useState('');
+  const [customText, setCustomText] = useState('');
+
+  const items = useMemo<WorkoutItem[]>(
+    () => [...(draft?.items ?? [])].sort((a, b) => a.createdAt - b.createdAt),
+    [draft?.items]
+  );
+
+  // Pause timer when finish sheet opens; resume when it closes
+  useEffect(() => {
+    if (finishOpen) pause();
+    else resume();
+  }, [finishOpen, pause, resume]);
+
+  return (
+    <SafeAreaView style={styles.root}>
+      <View style={styles.container}>
+        <TextInput
+          value={draft?.name ?? ''}
+          onChangeText={setDraftName}
+          placeholder="Write the name of your workout here"
+          placeholderTextColor="#71717a"
+          style={styles.nameInput}
+        />
+
+        <View style={styles.timerCard}>
+          <Text style={styles.timerText}>{mmss}</Text>
+        </View>
+
+        <View style={styles.actionsRow}>
+          <TouchableOpacity style={[styles.actionBtn, styles.btnBlue]} onPress={() => setExerciseOpen(true)} activeOpacity={0.9}>
+            <Text style={styles.actionTextWhite}>+ Exercise</Text>
+          </TouchableOpacity>
+          <TouchableOpacity style={[styles.actionBtn, styles.btnYellow]} onPress={() => setNoteOpen(true)} activeOpacity={0.9}>
+            <Text style={styles.actionTextDark}>+ Note</Text>
+          </TouchableOpacity>
+          <TouchableOpacity style={[styles.actionBtn, styles.btnGray]} onPress={() => setCustomOpen(true)} activeOpacity={0.9}>
+            <Text style={styles.actionTextWhite}>+ Custom</Text>
+          </TouchableOpacity>
+        </View>
+
+        <FlatList
+          data={items}
+          keyExtractor={(it) => it.id}
+          contentContainerStyle={{ paddingVertical: 12, gap: 10 }}
+          renderItem={({ item }) => (
+            <View style={styles.card}>
+              <View style={styles.cardHeader}>
+                <Text style={styles.cardType}>{item.type.toUpperCase()}</Text>
+                <Text style={styles.cardTime}>
+                  {new Date(item.createdAt).toLocaleTimeString()}
+                </Text>
+              </View>
+              {'name' in item && <Text style={styles.cardText}>{item.name}</Text>}
+              {'text' in item && <Text style={styles.cardText}>{item.text}</Text>}
+            </View>
+          )}
+          ListEmptyComponent={<Text style={styles.empty}>Add something to get started.</Text>}
+        />
+
+        <TouchableOpacity style={styles.finishBtn} onPress={() => setFinishOpen(true)} activeOpacity={0.95}>
+          <Text style={styles.finishText}>Save & Finish</Text>
+        </TouchableOpacity>
+      </View>
+
+      {/* Sheets */}
+      <Modal visible={exerciseOpen} onRequestClose={() => setExerciseOpen(false)} animationType="slide" presentationStyle="pageSheet">
+        <Sheet title="Add Exercise" onClose={() => setExerciseOpen(false)}>
+          <TextInput
+            value={exerciseName}
+            onChangeText={setExerciseName}
+            placeholder="Exercise name"
+            placeholderTextColor="#777"
+            style={styles.sheetInput}
+          />
+          <TouchableOpacity
+            style={styles.sheetPrimary}
+            onPress={() => {
+              const t = exerciseName.trim();
+              if (t) addExercise(t);
+              setExerciseName('');
+              setExerciseOpen(false);
+            }}
+          >
+            <Text style={styles.sheetPrimaryText}>Add Exercise</Text>
+          </TouchableOpacity>
+        </Sheet>
+      </Modal>
+
+      <Modal visible={noteOpen} onRequestClose={() => setNoteOpen(false)} animationType="slide" presentationStyle="pageSheet">
+        <Sheet title="Add Note" onClose={() => setNoteOpen(false)}>
+          <TextInput
+            value={noteText}
+            onChangeText={setNoteText}
+            placeholder="Type your note…"
+            placeholderTextColor="#777"
+            multiline
+            style={[styles.sheetInput, { minHeight: 120, textAlignVertical: 'top' }]}
+          />
+          <TouchableOpacity
+            style={styles.sheetPrimary}
+            onPress={() => {
+              const t = noteText.trim();
+              if (t) addNote(t);
+              setNoteText('');
+              setNoteOpen(false);
+            }}
+          >
+            <Text style={styles.sheetPrimaryText}>Save Note</Text>
+          </TouchableOpacity>
+        </Sheet>
+      </Modal>
+
+      <Modal visible={customOpen} onRequestClose={() => setCustomOpen(false)} animationType="slide" presentationStyle="pageSheet">
+        <Sheet title="Add Custom" onClose={() => setCustomOpen(false)}>
+          <TextInput
+            value={customText}
+            onChangeText={setCustomText}
+            placeholder="What do you want to log?"
+            placeholderTextColor="#777"
+            style={styles.sheetInput}
+          />
+          <TouchableOpacity
+            style={styles.sheetPrimary}
+            onPress={() => {
+              const t = customText.trim();
+              if (t) addCustom(t);
+              setCustomText('');
+              setCustomOpen(false);
+            }}
+          >
+            <Text style={styles.sheetPrimaryText}>Add Custom</Text>
+          </TouchableOpacity>
+        </Sheet>
+      </Modal>
+
+      {/* Finish (paused + JSON preview, saving disabled) */}
+      <Modal visible={finishOpen} onRequestClose={() => setFinishOpen(false)} animationType="slide" presentationStyle="pageSheet">
+        <Sheet title="Finish Workout (Preview)" onClose={() => setFinishOpen(false)}>
+          <Text style={styles.summary}>
+            {items.length} item(s) • Time {mmss} • (timer paused)
+          </Text>
+          <View style={styles.jsonBox}>
+            <ScrollView>
+              <Text style={styles.jsonText}>
+                {JSON.stringify({ draft, history }, null, 2)}
+              </Text>
+            </ScrollView>
+          </View>
+          <TouchableOpacity
+            style={[styles.finishConfirm, { opacity: 0.5 }]}
+            activeOpacity={1}
+            onPress={() => {
+              Alert.alert('Disabled', 'Saving is disabled for now. Close this sheet to resume the timer.');
+            }}
+          >
+            <Text style={styles.finishConfirmText}>Save Disabled</Text>
+          </TouchableOpacity>
+        </Sheet>
+      </Modal>
+    </SafeAreaView>
+  );
 }
 
+function Sheet({
+  title,
+  onClose,
+  children,
+}: {
+  title: string;
+  onClose: () => void;
+  children: React.ReactNode;
+}) {
+  return (
+    <SafeAreaView style={styles.sheetRoot}>
+      <View style={styles.grabber} />
+      <View style={styles.sheetHeader}>
+        <Text style={styles.sheetTitle}>{title}</Text>
+        <Pressable onPress={onClose} style={styles.closeBtn}>
+          <Text style={styles.closeText}>Close</Text>
+        </Pressable>
+      </View>
+      <View style={styles.sheetBody}>{children}</View>
+    </SafeAreaView>
+  );
+}
+
+const styles = StyleSheet.create({
+  root: { flex: 1, backgroundColor: '#0b0b0b' },
+  container: { flex: 1, padding: 16, gap: 12 },
+
+  nameInput: {
+    borderWidth: 1, borderColor: '#262626', backgroundColor: '#121212', color: 'white',
+    paddingHorizontal: 12, paddingVertical: 12, borderRadius: 12, fontSize: 16, fontWeight: '600',
+  },
+
+  timerCard: {
+    backgroundColor: '#111', borderWidth: 1, borderColor: '#222', borderRadius: 12,
+    paddingVertical: 14, alignItems: 'center',
+  },
+  timerText: { color: 'white', fontSize: 36, fontWeight: '800', letterSpacing: 1 },
+
+  actionsRow: { flexDirection: 'row', gap: 10 },
+  actionBtn: {
+    flex: 1,
+    paddingVertical: 14,
+    borderRadius: 12,
+    alignItems: 'center',
+    justifyContent: 'center',
+    borderWidth: 1,
+  },
+  btnBlue: { backgroundColor: '#0A84FF', borderColor: '#0A84FF' },
+  btnYellow: { backgroundColor: '#FFD60A', borderColor: '#FFD60A' },
+  btnGray: { backgroundColor: '#2b2b2b', borderColor: '#3a3a3a' },
+
+  actionTextWhite: { color: 'white', fontSize: 15, fontWeight: '800' },
+  actionTextDark: { color: '#111', fontSize: 15, fontWeight: '800' },
+
+  card: { backgroundColor: '#111', borderWidth: 1, borderColor: '#222', borderRadius: 12, padding: 12, gap: 6 },
+  cardHeader: { flexDirection: 'row', justifyContent: 'space-between' },
+  cardType: { color: '#a3a3a3', fontSize: 12, fontWeight: '700', letterSpacing: 0.5 },
+  cardTime: { color: '#6b7280', fontSize: 12 },
+  cardText: { color: 'white', fontSize: 15 },
+
+  empty: { color: '#6b7280', textAlign: 'center', marginTop: 8 },
+
+  finishBtn: { backgroundColor: '#22C55E', paddingVertical: 16, borderRadius: 12, alignItems: 'center' },
+  finishText: { color: 'white', fontSize: 16, fontWeight: '800' },
+
+  sheetRoot: { flex: 1, backgroundColor: '#0b0b0b' },
+  grabber: { alignSelf: 'center', width: 40, height: 5, borderRadius: 3, backgroundColor: '#333', marginTop: 8, marginBottom: 4 },
+  sheetHeader: { flexDirection: 'row', alignItems: 'center', paddingHorizontal: 16, paddingVertical: 8 },
+  sheetTitle: { color: 'white', fontSize: 18, fontWeight: '800', flex: 1 },
+  closeBtn: { paddingHorizontal: 12, paddingVertical: 8, borderRadius: 8, backgroundColor: '#151515' },
+  closeText: { color: '#0A84FF', fontWeight: '800' },
+  sheetBody: { flex: 1, padding: 16, gap: 12 },
+
+  sheetInput: {
+    borderWidth: 1, borderColor: '#333', backgroundColor: '#141414', color: 'white',
+    paddingHorizontal: 12, paddingVertical: 10, borderRadius: 10,
+  },
+  sheetPrimary: { backgroundColor: '#0A84FF', paddingVertical: 14, borderRadius: 10, alignItems: 'center' },
+  sheetPrimaryText: { color: 'white', fontSize: 16, fontWeight: '800' },
+
+  summary: { color: '#d1d5db', marginBottom: 12, fontSize: 15 },
+
+  jsonBox: {
+    flex: 1,
+    borderWidth: 1,
+    borderColor: '#262626',
+    backgroundColor: '#0f0f0f',
+    borderRadius: 10,
+    padding: 10,
+    marginBottom: 12,
+  },
+  jsonText: { color: '#cbd5e1', fontFamily: Platform.select({ ios: 'Menlo', android: 'monospace' }) as any, fontSize: 12 },
+
+  finishConfirm: { backgroundColor: '#22C55E', paddingVertical: 14, borderRadius: 10, alignItems: 'center' },
+  finishConfirmText: { color: 'white', fontSize: 16, fontWeight: '800' },
+});
